optimize whole backtesting system workflow

+++ bybit/BybitResultManager.py
@@ -9,18 +9,27 @@ import shutil
 import time
 
 class BybitResultManager:
-    def __init__(self, strategy, category, interval):
-        self.asset      = "Cryptocurrency"
+    def __init__(self, strategy, category, interval, delete_file):
+        self.asset      = "cryptocurrency"
         self.exchange   = "bybit"
         self.strategy   = strategy
         self.category   = category
         self.interval   = interval
         self.price_func = "kline"
 
+        self.set = "backtest_set"
+        self.delete_file = delete_file
+
     def _delete_useless_backtest_result(self):
         full_symbol_result_df = self._get_full_symbol_result_df()
         delete_list           = self._get_delete_list(full_symbol_result_df)
-        self._delete_folders(delete_list)
+
+        if self.delete_file == True:
+            self._delete_folders(delete_list)
+            print(f"{self.exchange}丨{self.strategy}丨{self.category}丨{self.interval}丨action = deleted useless backtest file")
+
+        else:
+            print(f"{self.exchange}丨{self.strategy}丨{self.category}丨{self.interval}丨action = generated useless backtest file_list")
 
     def _get_full_symbol_result_path(self):
         full_symbol_result_path = f"D:/backtest/{self.asset}/{self.exchange}/{self.strategy}/{self.category}/{self.interval}"
@@ -29,7 +38,7 @@ class BybitResultManager:
 
     def _get_full_symbol_result_df(self):
         full_symbol_result_path = self._get_full_symbol_result_path()
-        full_symbol_result_csv  = f"{full_symbol_result_path}/backtest_set/full_result/full_symbol_backtest_result.csv"
+        full_symbol_result_csv  = f"{full_symbol_result_path}/full_symbol_backtest_result.csv"
 
         full_symbol_result_df = pd.read_csv(full_symbol_result_csv)
